Update to session auth

+++ app/components/GradeBoardComponent.tsx
@@ -5,8 +5,7 @@ import styles from "./ComponentStyle.module.css";
 import WordComponent from "./WordComponent";
 import KeyBoardComponent from "./KeyBoardComponent";
 import ToolBoxComponent from "./ToolBoxComponent";
-import { useSession } from "next-auth/react"
-import { useRouter } from 'next/navigation'
+import { useCustomAuth } from "@/app/context/CustomAuthProvider";
 import StarComponent from "./StarComponent";
 
 const selectItems = [
@@ -38,11 +37,7 @@ const BoardComponent: React.FC<{}> = () => {
     const [marked, setMarked] = useState<boolean>(false);
     const [isOpen, setIsOpen] = useState(false);
     const [isLoading, setIsLoading] = useState(false);
-    const router = useRouter()
-
-    const { data: session, update } = useSession({
-        required: false,
-    })
+    const { isAuthenticated, login } = useCustomAuth();
 
     useEffect(() => {
         if (wordList.length == 0 && !isLoading) {
@@ -110,7 +105,7 @@ const BoardComponent: React.FC<{}> = () => {
 
     function fetchMarkList(wordList: Word[]) {
 
-        if (!session?.user) {
+        if (!isAuthenticated) {
             return wordList
         }
 
@@ -142,12 +137,10 @@ const BoardComponent: React.FC<{}> = () => {
     }
 
     function markWord() {
-
-        // TODO
-        // if (!session?.user) {
-        //     console.log("22222222222")
-        //     router.push('/api/auth/signin')
-        // }
+        if (!isAuthenticated) {
+            login()
+            return
+        }
 
         if (word) {
             fetch("/hts/api/v1/mark?word=" + word.word, { method: 'POST', }).then((response: Response) => {
@@ -159,12 +152,10 @@ const BoardComponent: React.FC<{}> = () => {
     }
 
     function unmarkWord() {
-
-        // TODO
-        // if (!session?.user) {
-        //     console.log("111111111")
-        //     router.push('/api/auth/signin')
-        // }
+        if (!isAuthenticated) {
+            login()
+            return
+        }
 
         if (word) {
             fetch("/hts/api/v1/mark?word=" + word.word, { method: 'DELETE', }).then((response: Response) => {
